Rename PeopleActionFactory to PeopleActions
